+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 CHANGELOG
 ===
-## 0.1.0
+## v0.2.0
+* New pipe - `Replace`
+* New filter - `Unique`
+* New feature `storage` with dependencies `pickleDb` and `md5`
+## v0.1.0
 * Initial version